requirements.txt
pyttsx3>=2.90
opencv-python>=4.5.0
face-recognition>=1.3.0
imutils>=0.5.4
requests>=2.25.0
cvzone>=1.5.0
numpy>=1.19.0
PyYAML>=6.0 